DolphinNoGUI: upgrade verify concept to generic disc image tool
