+++ Source/Core/DolphinNoGUI/CMakeLists.txt
@@ -5,11 +5,11 @@ add_executable(dolphin-nogui
   MainNoGUI.cpp
 )
 
-add_executable(dolphin-discimageverify
+add_executable(dolphin-discimagetool
   Platform.cpp
   Platform.h
   PlatformHeadless.cpp
-  MainDiscImageVerify.cpp
+  MainDiscImageTool.cpp
 )
 
 if(ENABLE_X11 AND X11_FOUND)
@@ -25,7 +25,7 @@ if(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
 endif()
 
 set_target_properties(dolphin-nogui PROPERTIES OUTPUT_NAME dolphin-emu-nogui)
-set_target_properties(dolphin-discimageverify PROPERTIES OUTPUT_NAME dolphin-emu-discimageverify)
+set_target_properties(dolphin-discimagetool PROPERTIES OUTPUT_NAME dolphin-emu-discimagetool)
 
 target_link_libraries(dolphin-nogui
 PRIVATE
@@ -34,7 +34,7 @@ PRIVATE
   cpp-optparse
 )
 
-target_link_libraries(dolphin-discimageverify
+target_link_libraries(dolphin-discimagetool
 PRIVATE
   core
   cpp-optparse
@@ -45,7 +45,7 @@ if(USE_DISCORD_PRESENCE)
 endif()
 
 set(CPACK_PACKAGE_EXECUTABLES ${CPACK_PACKAGE_EXECUTABLES} dolphin-nogui)
-set(CPACK_PACKAGE_EXECUTABLES ${CPACK_PACKAGE_EXECUTABLES} dolphin-discimageverify)
+set(CPACK_PACKAGE_EXECUTABLES ${CPACK_PACKAGE_EXECUTABLES} dolphin-discimagetool)
 install(TARGETS dolphin-nogui RUNTIME DESTINATION ${bindir})
-install(TARGETS dolphin-discimageverify RUNTIME DESTINATION ${bindir})
+install(TARGETS dolphin-discimagetool RUNTIME DESTINATION ${bindir})
